// src/shared/api/generated/models/supportInternalNoteAttachmentStatusResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportInternalNoteAttachmentStatusResponseDtoScanState } from "./supportInternalNoteAttachmentStatusResponseDtoScanState";
import type { SupportInternalNoteAttachmentStatusResponseDtoState } from "./supportInternalNoteAttachmentStatusResponseDtoState";

export interface SupportInternalNoteAttachmentStatusResponseDto {
  /** @nullable */
  attachedNoteId: string | null;
  canAttach: boolean;
  contentType: string;
  displayFilename: string;
  draftKey: string;
  /** @nullable */
  failureCode: string | null;
  id: string;
  scanState: SupportInternalNoteAttachmentStatusResponseDtoScanState;
  sizeBytes: number;
  state: SupportInternalNoteAttachmentStatusResponseDtoState;
  updatedAt: string;
}